requirements.txt
annotated-types==0.7.0
anyio==4.7.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.1.0
Flask-Uploads==0.2.1
fonttools==4.55.3
future==1.0.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
noisereduce==3.0.3
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
openai==1.58.1
opencv-python==4.7.0.72
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
